pyproject.toml
[project]
name = "my-fastapi-sqlite-books"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.116.2",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
]
